Renamed State as per Micheal's recommendation
