+++ doc/javadoc/com/CMPUT301W20T24/OnMyWay/State.html
@@ -98,7 +98,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.CMPUT301W20T24.OnMyWay.State</li>
+<li>com.CMPUT301W20T24.OnMyWay.UserRequestState</li>
 </ul>
 </li>
 </ul>
